requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.10.3
amqp==5.1.1
anyio==3.6.2
apache-airflow==2.5.3
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
apturl==0.5.2
argcomplete==3.0.5
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.1
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
billiard==4.1.0
bleach==5.0.1
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cachelib==0.9.0
cachetools==5.3.1
cattrs==22.2.0
celery==5.3.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudinary==1.30.0
colorama==0.4.3
colorlog==4.8.0
comm==0.1.2
command-not-found==0.3
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
croniter==1.3.14
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.13.0
django-db==0.0.7
django-filter==22.1
django-htmx==1.15.0
django-model-utils==4.3.1
django-recaptcha==3.0.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-sesame==3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.19
drf-yasg==1.21.5
duplicity==0.8.12.0
email-validator==1.3.1
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
findspark==2.0.1
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.39.0
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.0
httplib2==0.14.0
httpx==0.24.0
idna==2.8
importlib-metadata==4.13.0
importlib-resources==5.10.2
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
itypes==1.2.0
jax==0.4.11
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
keras==2.12.0
keyring==18.0.1
kiwisolver==1.4.4
kombu==5.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.9.0
libclang==16.0.0
linkify-it-py==2.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.1.0
monotonic==1.5
more-itertools==4.2.0
multidict==6.0.4
mxnet-mkl==1.6.0
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.1
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.3
pendulum==2.1.2
pexpect==4.6.0
phonenumbers==8.13.15
phonenumberslite==8.13.15
pickleshare==0.7.5
Pillow==7.0.0
pipdeptree==2.7.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.15.0
prompt-toolkit==3.0.38
protobuf==4.23.2
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.14.0
psutil==5.9.4
psycopg2==2.9.6
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.19.3
pyscopg2==66.0.2
pyspark==3.3.2
pytest==7.2.0
pytest-django==4.5.2
python-apt==2.0.1+ubuntu0.20.4.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.8
python-gnupg==0.4.5
python-json-logger==2.0.4
python-keyczar==0.716
python-nvd3==0.15.0
python-slugify==8.0.1
python3-openid==3.2.0
pythondialog==3.4.0
pytz==2022.7.1
pytzdata==2020.1
pyxdg==0.26
PyYAML==6.0
pyzmq==24.0.1
reportlab==3.5.34
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rich==13.3.4
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.8.0
sesame==0.3.3
setproctitle==1.3.2
simplejson==3.16.0
sip==4.19.21
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.0
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
texttable==1.6.2
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.8.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.1
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.13
usb-creator==0.3.7
utils==1.0.1
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
whitenoise==6.3.0
wrapt==1.14.1
WTForms==3.0.1
xkit==0.0.0
yarl==1.8.2
zipp==3.11.0
